+++ dynamic-simulation/dynamic-simulation-api/src/main/java/com/powsybl/dynamicsimulation/DynamicSimulationParameters.java
@@ -37,8 +37,8 @@ public class DynamicSimulationParameters extends AbstractExtendable<DynamicSimul
 
     public static final String VERSION = "1.0";
 
-    public static final int DEFAULT_START_TIME = 0;
-    public static final int DEFAULT_STOP_TIME = 10;
+    public static final double DEFAULT_START_TIME = 0d;
+    public static final double DEFAULT_STOP_TIME = 10d;
 
     private static final Supplier<ExtensionProviders<ConfigLoader>> SUPPLIER = Suppliers
         .memoize(() -> ExtensionProviders.createProvider(ConfigLoader.class, "dynamic-simulation-parameters"));
@@ -65,14 +65,14 @@ public class DynamicSimulationParameters extends AbstractExtendable<DynamicSimul
 
         platformConfig.getOptionalModuleConfig("dynamic-simulation-default-parameters")
             .ifPresent(config -> {
-                parameters.setStartTime(config.getIntProperty("startTime", DEFAULT_START_TIME));
-                parameters.setStopTime(config.getIntProperty("stopTime", DEFAULT_STOP_TIME));
+                parameters.setStartTime(config.getDoubleProperty("startTime", DEFAULT_START_TIME));
+                parameters.setStopTime(config.getDoubleProperty("stopTime", DEFAULT_STOP_TIME));
             });
     }
 
-    private int startTime;
+    private double startTime;
 
-    private int stopTime;
+    private double stopTime;
 
     /**
      * Constructor with given parameters
@@ -82,7 +82,7 @@ public class DynamicSimulationParameters extends AbstractExtendable<DynamicSimul
      * @param stopTime  instant of time at which the dynamic simulation ends, in
      *                  seconds
      */
-    public DynamicSimulationParameters(int startTime, int stopTime) {
+    public DynamicSimulationParameters(double startTime, double stopTime) {
         if (startTime < 0) {
             throw new IllegalStateException("Start time should be zero or positive");
         }
@@ -103,7 +103,7 @@ public class DynamicSimulationParameters extends AbstractExtendable<DynamicSimul
         stopTime = other.stopTime;
     }
 
-    public int getStartTime() {
+    public double getStartTime() {
         return startTime;
     }
 
@@ -113,7 +113,7 @@ public class DynamicSimulationParameters extends AbstractExtendable<DynamicSimul
      *                  seconds
      * @return
      */
-    public DynamicSimulationParameters setStartTime(int startTime) {
+    public DynamicSimulationParameters setStartTime(double startTime) {
         if (startTime < 0) {
             throw new IllegalStateException("Start time should be zero or positive");
         }
@@ -121,7 +121,7 @@ public class DynamicSimulationParameters extends AbstractExtendable<DynamicSimul
         return this;
     }
 
-    public int getStopTime() {
+    public double getStopTime() {
         return stopTime;
     }
 
@@ -131,7 +131,7 @@ public class DynamicSimulationParameters extends AbstractExtendable<DynamicSimul
      *                 seconds
      * @return
      */
-    public DynamicSimulationParameters setStopTime(int stopTime) {
+    public DynamicSimulationParameters setStopTime(double stopTime) {
         if (stopTime <= startTime) {
             throw new IllegalStateException("Stop time should be greater than start time");
         }
